requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
drf-yasg>=1.21
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
dj-database-url>=2.1.0
